fix default redirect
